Webinar template Changes
