+++ eds/blocks/columns/columns.css
@@ -50,6 +50,10 @@
   margin-bottom: 0.5rem
 }
 
+.columns-wrapper :is(.mt-8) {
+  margin-top: 2rem
+}
+
 .columns-wrapper :is(.block) {
   display: block
 }
@@ -78,8 +82,25 @@
   flex-direction: column
 }
 
+.columns-wrapper :is(.items-center) {
+  align-items: center
+}
+
+.columns-wrapper :is(.justify-between) {
+  justify-content: space-between
+}
+
 .columns-wrapper :is(.gap-y-6) {
   row-gap: 1.5rem
+}
+
+.columns-wrapper :is(.rounded) {
+  border-radius: 0.25rem
+}
+
+.columns-wrapper :is(.bg-\[\#EDF6F7\]) {
+  --tw-bg-opacity: 1;
+  background-color: rgb(237 246 247 / var(--tw-bg-opacity))
 }
 
 .columns-wrapper :is(.object-cover) {
@@ -92,6 +113,20 @@
   padding-right: 1.5rem
 }
 
+.columns-wrapper :is(.px-2) {
+  padding-left: 0.5rem;
+  padding-right: 0.5rem
+}
+
+.columns-wrapper :is(.py-1) {
+  padding-top: 0.25rem;
+  padding-bottom: 0.25rem
+}
+
+.columns-wrapper :is(.pt-4) {
+  padding-top: 1rem
+}
+
 .columns-wrapper :is(.font-\[rockwell\]) {
   font-family: rockwell
 }
@@ -99,6 +134,19 @@
 .columns-wrapper :is(.text-sm) {
   font-size: 0.875rem;
   line-height: 1.25rem
+}
+
+.columns-wrapper :is(.text-xs) {
+  font-size: 0.75rem;
+  line-height: 1rem
+}
+
+.columns-wrapper :is(.font-semibold) {
+  font-weight: 600
+}
+
+.columns-wrapper :is(.font-normal) {
+  font-weight: 400
 }
 
 .columns-wrapper :is(.capitalize) {
@@ -114,7 +162,25 @@
   color: rgb(67 86 86 / var(--tw-text-opacity))
 }
 
+.columns-wrapper :is(.text-\[\#378189\]) {
+  --tw-text-opacity: 1;
+  color: rgb(55 129 137 / var(--tw-text-opacity))
+}
+
+.columns-wrapper :is(.text-\[\#8B8B8B\]) {
+  --tw-text-opacity: 1;
+  color: rgb(139 139 139 / var(--tw-text-opacity))
+}
+
 @media (min-width: 768px) {
+  .columns-wrapper :is(.md\:mt-12) {
+    margin-top: 3rem
+  }
+
+  .columns-wrapper :is(.md\:inline-flex) {
+    display: inline-flex
+  }
+
   .columns-wrapper :is(.md\:flex-row) {
     flex-direction: row
   }
