+++ eds/templates/webinar/webinar.css
@@ -18,6 +18,10 @@
   align-items: center
 }
 
+.justify-center {
+  justify-content: center
+}
+
 .justify-between {
   justify-content: space-between
 }
